Moved consumer e-service details alert into a separate component
