+++ src/pages/ConsumerEServiceDetailsPage/ConsumerEServiceDetails.page.tsx
@@ -1,14 +1,13 @@
+import React from 'react'
 import { EServiceQueries } from '@/api/eservice'
 import { PageBottomActionsContainer, PageContainer } from '@/components/layout/containers'
 import { PageContainerSkeleton } from '@/components/layout/containers/PageContainer'
 import { EServiceDetails, EServiceDetailsSkeleton } from '@/components/shared/EServiceDetails'
 import useGetEServiceConsumerActions from '@/hooks/useGetEServiceConsumerActions'
 import { Link, useParams } from '@/router'
-import { checkIfAlreadySubscribed, checkIfhasAlreadyAgreementDraft } from '@/utils/agreement.utils'
 import { formatTopSideActions } from '@/utils/common.utils'
-import { Alert, Stack } from '@mui/material'
-import React from 'react'
 import { useTranslation } from 'react-i18next'
+import { ConsumerEServiceDetailsAlerts } from './components/ConsumerEServiceDetailsAlerts'
 
 const ConsumerEServiceDetailsPage: React.FC = () => {
   return (
@@ -25,37 +24,15 @@ const ConsumerEServiceDetailsPageContent: React.FC = () => {
   const { data: descriptor } = EServiceQueries.useGetDescriptorCatalog(eserviceId, descriptorId)
   const { actions } = useGetEServiceConsumerActions(descriptor?.eservice, descriptor)
 
-  const isMine = descriptor?.eservice.isMine
-  const isSubscribed = checkIfAlreadySubscribed(descriptor?.eservice)
-  const hasAgreementDraft = checkIfhasAlreadyAgreementDraft(descriptor?.eservice)
-
   const topSideActions = formatTopSideActions(actions)
-
-  // Only show missing certified attributes alert when...
-  const shouldShowMissingCertifiedAttributesAlert =
-    // ...the e-service is not owned by the active party...
-    !isMine &&
-    // ... the party doesn't own all the certified attributes required...
-    !descriptor?.eservice.hasCertifiedAttributes &&
-    // ... the e-service'slatest active descriptor is the actual descriptor the user is viewing...
-    descriptor?.eservice.activeDescriptor?.id === descriptor?.id &&
-    /// ... and it is not archived.
-    descriptor?.state !== 'ARCHIVED'
 
   return (
     <PageContainer
       title={descriptor?.eservice.name || ''}
-      description={descriptor?.eservice.name}
+      description={descriptor?.eservice.description || ''}
       topSideActions={topSideActions}
     >
-      <Stack spacing={2}>
-        {isMine && <Alert severity="info">{t('read.alert.youAreTheProvider')}</Alert>}
-        {shouldShowMissingCertifiedAttributesAlert && (
-          <Alert severity="info">{t('read.alert.missingCertifiedAttributes')}</Alert>
-        )}
-        {isSubscribed && <Alert severity="info">{t('read.alert.alreadySubscribed')}</Alert>}
-        {hasAgreementDraft && <Alert severity="info">{t('read.alert.hasAgreementDraft')}</Alert>}
-      </Stack>
+      <ConsumerEServiceDetailsAlerts descriptor={descriptor} />
 
       {descriptor && <EServiceDetails descriptor={descriptor} />}
       {!descriptor && <EServiceDetailsSkeleton />}
